+++ R/SplitFunctions.R
@@ -229,6 +229,10 @@ TipLabels.matrix <- function (x) colnames(x)
 TipLabels.phylo <- function (x) x$tip.label
 
 #' @rdname TipLabels
+#' @export
+TipLabels.TreeNumber <- function (x) x$tip.label
+
+#' @rdname TipLabels
 #' @family Splits operations
 #' @export
 TipLabels.Splits <- function (x) attr(x, 'tip.label')