+++ R/zzz.R
@@ -5,6 +5,7 @@
 ## Reminders when releasing for CRAN
 release_questions <- function() {
   c(
+    "Have you really updated codemeta.json with codemetar::write_codemeta()?",
     "Have you cleared GitHub issues for this release milestone?",
     "Is the code free of #TODOs?",
     "Have you checked the Vignettes for sanity?",